feat(planner): Add planner project as a git submodule
This commit introduces the `planner` application as a git submodule to integrate its functionality into the main project.

The submodule is sourced from the repository at `https://github.com/rbrtpv/planner.git` and is added at the path `react-root/react-dev/planner`. Using a submodule allows for independent development and versioning of the planner project while keeping it synchronized with this repository.